feat(database): use abstract base class for db operations
